Implemented AABB Collision Detection
Each tile and the Player now have FloatRects attached to their positions. For checking collisions, we can conveniently use a built in SFML function intersects(FloatRect), which will return true if the Rects are overlapping.

This also makes cursor programming easier, because of another function contains(Vector2f), which returns true if the point is within the Rect.

The Origins of all entities has been reset to their (0,0)'s rather than their midpoints, except the Player where it's offset, since the Player's Rect is smaller than the player (this is so that you can easily fit into 1-tile-wide spaces.)